trySetPieceInEnableOffset and toString
toString : return concatened abbreviation and it offset for each piece in board. This turns simpler to compare similar configurations.

trySetPieceInEnableOffset : call canSetPiece and SetPiece, but iterate for all offset after the start one.

use row and column base on zero facilitate to calculete them based on offset.
See (http://stackoverflow.com/questions/11821899/how-to-get-row-and-column-from-index)

important start from initial offset to not get an error with it own position.